+++ src/db/index.ts
@@ -1,2 +1,2 @@
 export {DB, DBOptions} from "./interface";
-export {LevelDB, LevelDBOptions} from "./level";
+export {LevelDB, LevelDBOptions} from "./impl/level";
